requirements.txt
aiogram==2.15
python-dotenv==0.19.0


